Add index.html Add styles.css Add nav bar Add links Add header Add button Add divs Add img
Added in the top portion of the website with html and styled it with css.
Working in sections makes the website easier to manage.
Writing out your html depending on how want it displayed is important for styling.
Properly naming divs with a class and or id allows for you to style with ease.